Move creator auth models to appropriate packages
