+++ src/main/kotlin/com/appifyhub/monolith/network/creator/user/ops/ApiKeyRequest.kt
@@ -1,4 +1,4 @@
-package com.appifyhub.monolith.network.auth
+package com.appifyhub.monolith.network.creator.user.ops
 
 import com.fasterxml.jackson.annotation.JsonInclude
 import com.fasterxml.jackson.annotation.JsonInclude.Include
@@ -6,8 +6,6 @@ import com.fasterxml.jackson.annotation.JsonProperty
 import java.io.Serializable
 
 @JsonInclude(Include.NON_NULL)
-data class CreatorCredentialsRequest(
-  @JsonProperty("universal_id") val universalId: String,
-  @JsonProperty("signature") val signature: String,
+data class ApiKeyRequest(
   @JsonProperty("origin") val origin: String? = null,
 ) : Serializable
